[youtube] Bump version and update simpleText usage
This updates any remaining uses of the `simpleText` key to use the
`textFrom` helper to prevent future issues if any other values switch
over to the new list-based `runs` format.

`subtext` is also changed to a list so we only have to append the actual
values and can join with pipe separators when it's passed to the Item.